% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{annotateBedWithGenes}
\alias{annotateBedWithGenes}
\title{Annotate bed regions with genes}
\usage{
annotateBedWithGenes(bed_table, proteinCodingOnly = TRUE, genomev = "hg19")
}
\arguments{
\item{bed_table}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique.}

\item{proteinCodingOnly}{if TRUE then annotate only protein coding genes}

\item{genomev}{genome version, hg19 or hg38}
}
\value{
annotated bed_table
}
\description{
Given a table of bed regions, annotate each regions with the genes that overlap them.
}
