% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bedpeBreakpointsToPositions}
\alias{bedpeBreakpointsToPositions}
\title{Extract breakpoint positions from a bedpe structural variants table}
\usage{
bedpeBreakpointsToPositions(sv_bedpe, copycolumns = NULL)
}
\arguments{
\item{sv_bedpe}{data frame with required columns: chr1, start1, end1, chr2, start2, end2}

\item{copycolumns}{names of sv_bedpe columns to copy, for example to carry over
the mutation type or id of the structural variants}
}
\value{
positions of the breakpoints
}
\description{
Given a table containing structural variants in bedpe format, with two break
points in each row, return all the breakpoints as positions. For eacn breakpoint,
the corresponding position will be the floor of the average of start and end positions.
}
