% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlatePositionsWithBedRegions}
\alias{correlatePositionsWithBedRegions}
\title{Correlate positions with bed regions}
\usage{
correlatePositionsWithBedRegions(
  positions,
  bed_table,
  nsamples = 0,
  altHypothesis = "greaterthan",
  resamplePositionsFlag = TRUE,
  resampleBedRegionsFlag = TRUE,
  resampleBedRegionsAllowOverlap = TRUE,
  genomev = "hg19",
  samplingRegions = NULL,
  randomSeed = NULL,
  nparallel = 1
)
}
\arguments{
\item{positions}{data frame with required columns: chr, position, id}

\item{nsamples}{number of resampling used to determine the NULL distribution}

\item{altHypothesis}{greatherthan (the default) or lowerthan can be used}

\item{resamplePositionsFlag}{if TRUE then positions will be resampled to determine the NULL distribution. Note that at least one between resamplePositionsFlag and resampleBedRegionsFlag should be TRUE}

\item{resampleBedRegionsFlag}{if TRUE then bed_table will be resampled to determine the NULL distribution. Note that at least one between resamplePositionsFlag and resampleBedRegionsFlag should be TRUE}

\item{resampleBedRegionsAllowOverlap}{allow overlapping bed regions when resampling bed_table}

\item{genomev}{hg19 or hg38}

\item{samplingRegions}{supply your own sampling regions for resampling positions and/or bed_table.}

\item{randomSeed}{set a random seed for the resampling}

\item{nparallel}{how many parallel processes to use when running the resampling}

\item{bed_regions}{data frame with required columns: chr, start, end, id}
}
\value{
data frame of ordered positions
}
\description{
Given a set of positions and a set of regions, calculate the overlap between
the positions and the regions (see the function intersectPositionsAndBedRegions).
If nsamples is greater than zero, then positions and/or regions are resampled
randomly so that a NULL distribution of overlaps can be obtained. Notice that
the the bigger nsamples is, the more accurate the p-value for rejecting the NULL
hypothesis will be. A returned p-value of 0 just means that the p-value reached
the limit of decimal numbers that can be obtained and should be interpreted as
"< 1/nsamples".
}
