% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{distanceOfPositionToNearestBedRegion}
\alias{distanceOfPositionToNearestBedRegion}
\title{Compute distance of positions to the nearest bed region}
\usage{
distanceOfPositionToNearestBedRegion(positions, bed_table)
}
\arguments{
\item{positions}{data frame with required columns: chr, position, id}

\item{bed_table}{data frame with required columns: chr, start, end, id}
}
\value{
annotated positions with distance to nearest bed region
}
\description{
Given a set of positions and a set of bed regions, annotate the distance of
each position to the nearest bed region. Distance is 0 if the position is in
a region, and NA if the position is on a chromosome where there are no regions.
Regions need to be non-overlapping.
}
