% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{extendBedRegions}
\alias{extendBedRegions}
\title{Extend bed regions}
\usage{
extendBedRegions(bed_table, extended)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end}

\item{extended}{length in number of bases by which each region should be
extended in both directions}
}
\value{
updated bed_table
}
\description{
Given a table of bed regions, extend the regions in both directions by a given
length. When extending at the beginning of a chromosome, 1 will be
the minimum position. This function will extend bed regions allowing bed regions
to overlap. Use function extendBedRegionsWithNoOverlap for extending non-overlapping
regions and preserving the non-overlap property.
}
