% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{extendBedRegionsWithNoOverlap}
\alias{extendBedRegionsWithNoOverlap}
\title{Extend bed regions preserving no overlap}
\usage{
extendBedRegionsWithNoOverlap(bed_table, extended)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end}

\item{extended}{length in number of bases by which each region should be
extended in both directions}
}
\value{
updated bed_table
}
\description{
Given a table of non-overlapping bed regions, extend the regions
in both directions by a given length. If two regions are next to each
other and are at a distance that is less than 2 times the extend length,
then they will be both extended by half such distance so that they become
adjacent regions. When extending at the beginning of a chromosome, 1 will be
the minimum position.
}
