% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowdata.R
\name{formatBedDataToWindow}
\alias{formatBedDataToWindow}
\title{Format bed data to window}
\usage{
formatBedDataToWindow(bed_table, genomev = "hg19", windowSize = 5e+05)
}
\arguments{
\item{bed_table}{data frame with at least three required columns: chr, start, end, and optionally signal}

\item{genomev}{genome version to use, hg19 or hg38}

\item{windowSize}{window size to use. The last window at the end of each chromosome is
likely to be shorter so that the final window does not cover coordinates that are outside the genome.}
}
\value{
data frame of aggregated bed region signal for each window across the whole genome
}
\description{
Segment a genome into windows of a fixed size and for each window find
a set of overlapping bed regions and calculate an aggregated signal.
If the bed regions in the bed_table do not have a signal column, then
the aggregated signal for a window will simply be the proportion of the window
that is covered by the bed regions. This is equivalent to specifying a signal
of 1 for all bed regions. If a signal column is specified, then the aggregated
signal will be a weighted sum, where the signal of each region is multiplied
by the proportion of the window covered by that region.
}
