% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowdata.R
\name{formatBedpeBreakpointsDataToWindow}
\alias{formatBedpeBreakpointsDataToWindow}
\title{Format structural variant breakpoint position data to window}
\usage{
formatBedpeBreakpointsDataToWindow(
  sv_bedpe,
  genomev = "hg19",
  windowSize = 5e+05
)
}
\arguments{
\item{sv_bedpe}{data.frame with required columns: chrom1, start1, end1, chrom2, start2, end2}

\item{genomev}{genome version to use, hg19 or hg38}

\item{windowSize}{window size to use. The last window at the end of each chromosome is
likely to be shorter so that the final window does not cover coordinates that are outside the genome.}
}
\value{
data frame of counts of positions in windows across the whole genome.
Returns only the windows with non-zero data
}
\description{
Segment a genome into windows of a fixed size and count how many structural
variant breakpoints are in each window.
}
