% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowdata.R
\name{formatPositionDataToWindow}
\alias{formatPositionDataToWindow}
\title{Format positions data to window}
\usage{
formatPositionDataToWindow(positions, genomev = "hg19", windowSize = 5e+05)
}
\arguments{
\item{positions}{data frame with at least two required columns: chr, position}

\item{genomev}{genome version to use, hg19 or hg38}

\item{windowSize}{window size to use. The last window at the end of each chromosome is
likely to be shorter so that the final window does not cover coordinates that are outside the genome.}
}
\value{
data frame of counts of positions in windows across the whole genome.
Returns only the windows with non-zero data
}
\description{
Segment a genome into windows of a fixed size and count how many positions are
in each window.
}
