% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowdata.R
\name{getBedFromWindowData}
\alias{getBedFromWindowData}
\title{Get bed table from window data}
\usage{
getBedFromWindowData(windowData)
}
\arguments{
\item{windowData}{data.frame where the row names have the format chr_start_end}
}
\value{
bed data frame
}
\description{
A data frame containing window data has the window coordinates as row names.
This function can be used to extract the window coordinates from the row names
and add them to the table as chr, start, end columns.
}
