% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{getChromosomesBedTable}
\alias{getChromosomesBedTable}
\title{Get chromosomes bed table}
\usage{
getChromosomesBedTable(genomev)
}
\arguments{
\item{genomev}{hg19 or hg38}
}
\value{
chromosomes bed table
}
\description{
Return a bed table with the list of chromosomes.
}
