% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{getIMD}
\alias{getIMD}
\title{Compute inter-mutational distance of a set of positions}
\usage{
getIMD(positions)
}
\arguments{
\item{positions}{data frame with required columns: chr and position}
}
\value{
left, right and average inter-mutational distance
}
\description{
This function returns both the left and right IMD of each given position
as well as the average IMD.
}
