% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersectBed_nonOverlapping}
\alias{intersectBed_nonOverlapping}
\title{Intersect two sets of non-overlapping bed regions}
\usage{
intersectBed_nonOverlapping(bed_table1, bed_table2, computeStats = TRUE)
}
\arguments{
\item{bed_table1}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique. Regions cannot overlap.}

\item{bed_table2}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique. Regions cannot overlap.}

\item{computeStats}{if FALSE, intersect stats will not be calculated and only the id maps mapping
position ids to region ids and viceversa will be returned. This is meant to save compute time when
intersectBed_nonOverlapping is run inside intersectBed.}
}
\value{
object with details intersection statistics
}
\description{
Given two tables of bed regions, find the regions from the first table that
overlap the regions in the second table, and conversely the regions in the second
table that overlap the regions in the first table.
If the regions have classes, then find how many regions in the first table for each
class of regions are contained in each region or class of regions in the second table,
and find how many regions in the second table for each class of regions contain each
region or class of regions in the second table.
This function is restricted to only non-overlapping bed regions. For overlapping
bed regions, use the more general function intersectBed.
}
