% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersectionStats}
\alias{intersectionStats}
\title{Intersection statistics}
\usage{
intersectionStats(idMap, idclassmap1, idclassmap2)
}
\arguments{
\item{idMap}{list object where the names are ids of the first group, and each id maps to a vector of ids of the second group}

\item{idclassmap1}{data frame with required columns id and class, id must be unique}

\item{idclassmap2}{data frame with required columns id and class, id must be unique}
}
\value{
object with details intersection statistics
}
\description{
This is a low level function, the engine that calculates the overlap between
any two groups of genomics entities. Entities (e.g. positions and regions),
have ids and classes, so that ids are unique for a given group of entities,
while classes may be repeated, so that subgroups of entities can belong to
the same class. The intersectionStats function assumes that, whatever the
entities, the intersection of two groups of entities has been computed and
is represented by an idMap, which maps the ids from the first group to the
ids of the second group. The mapping is bidirectional, i.e. if an entity
in the first group intersects an entity in the second group then the converse
is also true. The mapping from id to class for each group needs to be provided
(idclassmap1 and idclassmap2), so that the statistics for the classes can be computed.
To obtain all statistics, one needs to run intersectionStats twice, swapping
the two groups, so that in the second run idMap maps from the ids of the second
group to the ids of the first group and idclassmap1 and idclassmap2 are swapped.
}
