% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersectionStatsComplete}
\alias{intersectionStatsComplete}
\title{Intersection statistics complete}
\usage{
intersectionStatsComplete(
  idMap1to2,
  idMap2to1 = NULL,
  idclassmap1,
  idclassmap2
)
}
\arguments{
\item{idMap1to2}{list object where the names are ids of the first group, and each id maps to a vector of ids of the second group}

\item{idMap2to1}{list object where the names are ids of the second group, and each id maps to a vector of ids of the first group.
If NULL, this will be computed automatically from idMap1to2}

\item{idclassmap1}{data frame with required columns id and class, id must be unique}

\item{idclassmap2}{data frame with required columns id and class, id must be unique}
}
\value{
object with details intersection statistics
}
\description{
This is a low level function, the engine that calculates the overlap between
any two groups of genomics entities. Entities (e.g. positions and regions),
have ids and classes, so that ids are unique for a given group of entities,
while classes may be repeated, so that subgroups of entities can belong to
the same class. The intersectionStats function assumes that, whatever the
entities, the intersection of two groups of entities has been computed and
is represented by an idMap1to2, which maps the ids from the first group to the
ids of the second group. The mapping is bidirectional, i.e. if an entity
in the first group intersects an entity in the second group then the converse
is also true. The converse idMap2to1 can also be provided, or it will be inferred
automatically if left NULL.
The mapping from id to class for each group needs to be provided
(idclassmap1 and idclassmap2), so that the statistics for the classes can be computed.
In practice, intersectionStatsComplete calls intersectionStats twice, so that
statistics of intersection can be calculated from both direction, i.e. counting elements of
of the first group that overlap the second group, or counting elements of the
second group that overlap the first group.
}
