% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersectPositionsAndBedRegions_nonOverlapping}
\alias{intersectPositionsAndBedRegions_nonOverlapping}
\title{Intersect positions with non-overlapping bed regions}
\usage{
intersectPositionsAndBedRegions_nonOverlapping(
  positions,
  bed_table,
  computeStats = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{positions}{data frame containing positions, with required columns chr, position, id and optionally class. Value in the id column must be unique}

\item{bed_table}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique. Regions cannot overlap.}

\item{computeStats}{if FALSE, intersect stats will not be calculated and only the id maps mapping
position ids to region ids and viceversa will be returned. This is meant to save compute time when
intersectPositionsAndBedRegions_nonOverlapping is run inside intersectPositionsAndBedRegions.}

\item{verbose}{set to FALSE to suppress the info messages. Warning and error messages will still be shown}
}
\value{
object with details intersection statistics
}
\description{
Given a table of positions and a table of bed regions, find the positions that
are contained in the regions, and conversely the regions that overlap given positions.
If positions and/or bed regions have classes, then find how many positions for each
class of positions are contained in each region or class of regions, and find how many regions for
each class of regions contain each position or each class of positions.
This function is restricted to only non-overlapping bed regions. For overlapping
bed regions, use the more general function intersectPositionsAndBedRegions.
}
