% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowdata.R
\name{mergeWindowData}
\alias{mergeWindowData}
\title{Merge window data}
\usage{
mergeWindowData(windowData1, windowData2, missingValueFill = NA)
}
\arguments{
\item{windowData1}{data.frame where the row names have the format chr_start_end}

\item{windowData2}{data.frame where the row names have the format chr_start_end}

\item{missingValueFill}{filler to use for the missing data, NA as default}
}
\value{
merged window data
}
\description{
Combine two data frames containing window data. The function expects that the
two data frames do not share column names, while they may share row names, where
the row names indicate the windows. The two data frame should use the same window
size. If a window is present in one data frame but missing in the second, a
missing value filler will be use, NA as default.
}
