% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{multipleCorrelations}
\alias{multipleCorrelations}
\title{Multiple correlations between positions and/or regions}
\usage{
multipleCorrelations(
  referenceEntities,
  compareEntitiesList,
  referenceEntitiesName = "referenceEntities",
  nsamples = 0,
  altHypothesis = "greaterthan",
  resampleCompareEntities = FALSE,
  resampleReferenceEntitiesAllowOverlap = TRUE,
  resampleCompareEntitiesAllowOverlap = TRUE,
  genomev = "hg19",
  samplingRegions = NULL,
  randomSeed = NULL,
  nparallel = 1
)
}
\arguments{
\item{referenceEntities}{data frame of positions or bed regions. If positions are used,
required columns are id, chr and position. If bed regions are used, required columns are
id, chr, start and end.}

\item{compareEntitiesList}{list of data frames, with each data frame either a table of positions
or bed regions. The list names will be used as the name of each entities set}

\item{referenceEntitiesName}{name to be use when referring to the reference entities}

\item{nsamples}{number of resampling used to determine the NULL distribution}

\item{altHypothesis}{greatherthan (the default) or lowerthan can be used}

\item{resampleCompareEntities}{if TRUE the compare entities will be resampled}

\item{resampleReferenceEntitiesAllowOverlap}{if TRUE and if the reference entities are
bed regions, the resampling will allow bed regions overlap}

\item{resampleCompareEntitiesAllowOverlap}{if TRUE and if the compare entities are
bed regions, the resampling will allow bed regions overlap}

\item{genomev}{hg19 or hg38}

\item{samplingRegions}{supply your own sampling regions for resampling positions and/or bed_table.}

\item{randomSeed}{set a random seed for the resampling}

\item{nparallel}{how many parallel processes to use when running the resampling}
}
\value{
correlation statistics and annotations
}
\description{
Given a table with a set of reference entities, which are either positions or
bed regions, calculate the correlation with a list of other entities, which
again can be positions and/or regions. The type of the entities, either positions
or bed regions, will be determined automatically by checking the column names,
where positions should have column names chr and position, while bed regions
should have column names chr, start and end. Statistical significance of the
intersection between the reference entities and other entities sets will be
determined by resampling the reference entities, and optionally also the entities
to compare to.
}
