% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plotBrokenDownBedRegions}
\alias{plotBrokenDownBedRegions}
\title{Plot broken down bed regions}
\usage{
plotBrokenDownBedRegions(
  bed_table,
  brokenDown_bed_table = NULL,
  filename = NULL,
  chrom,
  pstart,
  pend,
  region_colour = "#0067a5",
  segments_colour = "#F38400"
)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end, id, and optionally signal}

\item{brokenDown_bed_table}{This should be the result of breakDownOverlappingBedRegions(bed_table,...),
in case this has already been computed. If left NULL (the default), breakDownOverlappingBedRegions(bed_table)
is called by this function with default parameters}

\item{filename}{optional file name to plot to file, use .pdf extension}

\item{chrom}{chromosome location to plot}

\item{pstart}{genomic start location to plot}

\item{pend}{genomic end location to plot}

\item{region_colour}{plot colour for regions in bed_table}

\item{segments_colour}{plot colour of non-overlapping segment regions in brokenDown_bed_table}
}
\value{
list object with the results of breakDownOverlappingBedRegions and assignBedRegionsToNonOverlappingSets function calls
}
\description{
Given a table with bed regions, plot the bed regions within a given genomic
location (chrom:pstart-pend) so that overlapping regions are shown on separate
lines. This plot function is useful to visualise overlapping bed regions.
This function calls breakDownOverlappingBedRegions (if brokenDown_bed_table is NULL)
and assignBedRegionsToNonOverlappingSets, and the results from these function
calls are returned in a return object list.
}
