% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{randomPositionInRegions}
\alias{randomPositionInRegions}
\title{Sample a random position from a set of regions}
\usage{
randomPositionInRegions(samplingRegions)
}
\arguments{
\item{samplingRegions}{data frame containing bed regions, with required columns chr, start, end, and optionally regionprob. If regionprob is missing, it will be proportional to the region size.}
}
\value{
sampled position
}
\description{
Given a table of bed regions, sample a random position.
}
