% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{resampleBedRegions}
\alias{resampleBedRegions}
\title{Resample bed regions}
\usage{
resampleBedRegions(
  bed_table,
  genomev = NULL,
  samplingRegions = NULL,
  randomSeed = NULL,
  allowRegionsOverlap = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{bed_table}{data.frame with positions, required columns are: chr, start, end}

\item{genomev}{reference genome version, hg19 or hg38. This parameter will be used only if
samplingRegions is NULL, to load the samplingRegions for the corresponding reference genome}

\item{samplingRegions}{data frame containing bed regions, with required columns chr, start, end,
and optionally regionprob. If regionprob is missing, it will be proportional to the region sizes
for each chromosome. If NULL, samplingRegions will be the set of regions that are mappable and
not blacklisted according to ENCODE for the reference genome specified by genomev}

\item{randomSeed}{set a random seed}

\item{verbose}{print additional output}
}
\value{
sampled regions
}
\description{
Given a table of bed regions, resample them to obtain a set of random bed regions.
The chromosome of each resampled region will be preserved. Regions will be resampled
within regions that are mappable and not blacklisted according to ENCODE.
Custom sampling regions can be specified with the parameter samplingRegions, which can
be used for example to assign custom probability to each region. Bear in mind that
for each region, the corresponding subsets of regions in the same chromosome will
be selected. By default (allowRegionsOverlap=TRUE) the resampled bed regions can
overlap. If allowRegionsOverlap is set the FALSE, then the resampled bed regions
are not allowed to overlap. This is implemented by updating the samplingRegions
table, subtracting each resampled region. The subtraction of a resampled region
from samplingRegions will split or reduce the size of a sampling region, so the
corresponding regionprob will be split or reduced proportionally to the size reduction.
}
