% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{resamplePositions}
\alias{resamplePositions}
\title{Resample a set of positions}
\usage{
resamplePositions(
  positions,
  genomev,
  samplingRegions = NULL,
  randomSeed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{positions}{data.frame with positions, required columns are: chr, position}

\item{genomev}{reference genome version, hg19 or hg38. This parameter will be used only if
samplingRegions is NULL, to load the samplingRegions for the corresponding reference genome}

\item{samplingRegions}{data frame containing bed regions, with required columns chr, start, end,
and optionally regionprob. If regionprob is missing, it will be proportional to the region sizes
for each chromosome. If NULL, samplingRegions will be the set of regions that are mappable and
not blacklisted according to ENCODE for the reference genome specified by genomev}

\item{randomSeed}{set a random seed}

\item{verbose}{print additional output}
}
\value{
sampled positions
}
\description{
Given a table of positions, resample them to obtain a set of random position.
The chromosome of each resampled position will be preserved. Positions will be resampled
within regions that are mappable and not blacklisted according to ENCODE.
Custom sampling regions can be specified by the parameter samplingRegions, which can
be used for example to assign custom probability to each region. Bear in mind that
for each position, the correspoinding subsets of regions in the same chromosome will
be selected, so it would be good practice to have the regionprob for each chromosome sum to 1.
}
