% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{resampleSV}
\alias{resampleSV}
\title{Resample structural variants}
\usage{
resampleSV(
  sv_bedpe,
  genomev,
  samplingRegions = NULL,
  randomSeed = NULL,
  maxattempts = 100,
  copyAdditionalColumns = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sv_bedpe}{data.frame with positions, required columns are: chrom1, start1, end1, chrom2, start2, end2, svclass}

\item{genomev}{reference genome version, hg19 or hg38. This parameter will be used only if
samplingRegions is NULL, to load the samplingRegions for the corresponding reference genome}

\item{samplingRegions}{data frame containing bed regions, with required columns chr, start, end,
and optionally regionprob. If regionprob is missing, it will be proportional to the region sizes
for each chromosome. If NULL, samplingRegions will be the set of regions that are mappable and
not blacklisted according to ENCODE for the reference genome specified by genomev}

\item{randomSeed}{set a random seed}

\item{maxattempts}{max number of attempts for resampling each SV. If maxattempts is reached, then the SV is skipped.}

\item{copyAdditionalColumns}{if TRUE, columns other than the required columns will be copied as well into the resampled SVs}

\item{verbose}{print additional output}
}
\value{
resampled SVs
}
\description{
Given a table of structural variants in bedpe format, resample them to obtain a set of random structural variants.
The chromosome of each resampled region will be preserved. Structural variant breakpoints will be resampled
within regions that are mappable and not blacklisted according to ENCODE.
Custom sampling regions can be specified with the parameter samplingRegions, which can
be used for example to assign custom probability to each region. Bear in mind that
for each region, the corresponding subsets of regions in the same chromosome will
be selected.
}
