% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sortBed}
\alias{sortBed}
\title{Sort bed}
\usage{
sortBed(
  bed_table,
  decreasingChr = FALSE,
  decreasingPos = FALSE,
  usePos = "start"
)
}
\arguments{
\item{bed_table}{data frame with at least three required columns: chr, start, end}

\item{decreasingChr}{if TRUE, sort so that the chromosomes are listed in reverse order}

\item{decreasingPos}{if TRUE, the bed regions in each chromosomes will be sorted from last to first}

\item{usePos}{which position to use for sorting, either start or end (default start)}
}
\value{
data frame of ordered bed regions
}
\description{
Sort a list of bed regions, according to the chromosome (chr column) and start
position (start column). If a bed region has end < start then the start and end
are swapped so that start <= end in the sorted bed table.
}
