% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sortPositions}
\alias{sortPositions}
\title{Sort positions}
\usage{
sortPositions(positions)
}
\arguments{
\item{positions}{data frame with at least two required columns: chr, position}
}
\value{
data frame of ordered positions
}
\description{
Sort a list of positions. Positions may be genomic coordinates of small variants or polymorphisms
}
