% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowdata.R
\name{sortWindowData}
\alias{sortWindowData}
\title{Sort window data}
\usage{
sortWindowData(windowData)
}
\arguments{
\item{windowData}{data.frame where the row names have the format chr_start_end}
}
\value{
sorted window data
}
\description{
Sort the rows of a data frame containing window data.
}
