% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{trimBlacklistedFromBed}
\alias{trimBlacklistedFromBed}
\title{Remove segments that overlap with blacklisted regions from a bed table}
\usage{
trimBlacklistedFromBed(bed_table, genomev, verbose = TRUE)
}
\arguments{
\item{bed_table}{data frame with required columns: id, chr, start, end, and optionally class}

\item{genomev}{hg19 or hg38}

\item{verbose}{set to FALSE to suppress the info messages. Warning and error messages will still be shown}
}
\value{
trimmed bed regions
}
\description{
Given a bed table with a set of regions, trim the regions to remove segments that
are blacklisted according to ENCODE or are outside the chromosome ranges.
Bear in mind that each region might be split into multiple regions and this will
results in multiple regions with the same value in the id column (and the class column if specified).
}
