% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{trimNfromBed}
\alias{trimNfromBed}
\title{Remove N from bed table}
\usage{
trimNfromBed(bed_table, genomev, verbose = FALSE)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end}

\item{genomev}{hg19 or hg38}

\item{verbose}{set to FALSE to suppress the info messages. Warning and error messages will still be shown}
}
\value{
updated bed_table
}
\description{
Given a table of bed regions and a reference genome, trim and split the bed
bed regions to remove reference genome N positions.
}
