% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{assignBedRegionsToNonOverlappingSets}
\alias{assignBedRegionsToNonOverlappingSets}
\title{Assign bed regions to non-overlapping sets}
\usage{
assignBedRegionsToNonOverlappingSets(
  bed_table,
  brokenDown_bed_table = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end, id, and optionally signal}

\item{brokenDown_bed_table}{This should be the result of breakDownOverlappingBedRegions(bed_table,...),
in case this has already been computed. If left NULL (the default), breakDownOverlappingBedRegions(bed_table)
is called by this function with default parameters}

\item{verbose}{set to FALSE to suppress the info messages. Warning and error messages will still be shown}
}
\value{
list object assigning a set number (integer) to each bed region id
}
\description{
Given a table with bed regions, assign each region to a set, so that each
set contains no overlapping bed regions. If the regions in the input bed_table
do not overlap, then they will be all assigned to the same set. For example,
given region id A, start 1, end 10, and region id B, start 5, end 15, the
function returns the list list(A=1,B=2).
}
