% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{breakDownOverlappingBedRegions}
\alias{breakDownOverlappingBedRegions}
\title{Break down overlapping bed regions}
\usage{
breakDownOverlappingBedRegions(
  bed_table,
  aggregateSignalMode = "mean",
  aggregateTextColumns = NULL
)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end, id, and optionally signal}

\item{aggregateSignalMode}{can be either: mean or sum, relevant only if the signal column is present}

\item{aggregateTextColumns}{is a list of columns of bed_table that should be aggregated}
}
\value{
updated bed_table with non-overlapping regions
}
\description{
If a bed_table contains overlapping bed regions, split the overlapping regions
and return a table of non-overlapping bed regions. An id column in the input
bed_table data frame is required to keep track of which returned region is derived
from which input region or regions. For example, given region id A, start 1, end 10,
and region id B, start 5, end 15, the returned table will contain three regions:
region id A, start 1, end 4, region id "A;B", start 5, end 10, and region id B,
start 11, end 15.
}
