% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{checkBedRegionsOverlap}
\alias{checkBedRegionsOverlap}
\title{Check existence of bed regions overlaps}
\usage{
checkBedRegionsOverlap(bed_table)
}
\arguments{
\item{bed_table}{data frame with at least three required columns: chr, start, end}
}
\value{
NULL if no overlaps are found, or a list of chromosome names where overlaps were found
}
\description{
Return NULL if all the bed regions in bed_table are disjoint, otherwise return
the list of chromosomes where the region overlaps occur.
}
