% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlateBedRegions}
\alias{correlateBedRegions}
\title{Correlate two sets of bed regions}
\usage{
correlateBedRegions(
  bed_table1,
  bed_table2,
  nsamples = 0,
  altHypothesis = "greaterthan",
  resampleBedRegions1Flag = TRUE,
  resampleBedRegions2Flag = TRUE,
  resampleBedRegions1AllowOverlap = TRUE,
  resampleBedRegions2AllowOverlap = TRUE,
  genomev = "hg19",
  samplingRegions = NULL,
  randomSeed = NULL,
  returnResampledBedRegions1 = FALSE,
  returnResampledBedRegions2 = FALSE,
  resampled_bed_regions1_list = NULL,
  resampled_bed_regions2_list = NULL,
  nparallel = 1,
  verbose = TRUE
)
}
\arguments{
\item{bed_table1}{data frame with required columns: chr, start, end, id}

\item{bed_table2}{data frame with required columns: chr, start, end, id}

\item{nsamples}{number of resampling used to determine the NULL distribution}

\item{altHypothesis}{greatherthan (the default) or lowerthan can be used}

\item{resampleBedRegions1Flag}{if TRUE then bed_table1 will be resampled to determine the NULL distribution. Note that at least one between resampleBedRegions1Flag and resampleBedRegions2Flag should be TRUE}

\item{resampleBedRegions2Flag}{if TRUE then bed_table2 will be resampled to determine the NULL distribution. Note that at least one between resampleBedRegions1Flag and resampleBedRegions2Flag should be TRUE}

\item{resampleBedRegions1AllowOverlap}{allow overlapping bed regions when resampling bed_table1}

\item{resampleBedRegions2AllowOverlap}{allow overlapping bed regions when resampling bed_table2}

\item{genomev}{hg19 or hg38}

\item{samplingRegions}{supply your own sampling regions for resampling positions and/or bed_table.}

\item{randomSeed}{set a random seed for the resampling}

\item{returnResampledBedRegions1}{if TRUE and if resampling of bed_table1 was performed, return a list with all the bed_table1 data frames resampled}

\item{returnResampledBedRegions2}{if TRUE and if resampling of bed_table2 was performed, return a list with all the bed_table2 data frames resampled}

\item{resampled_bed_regions1_list}{supply your own resampled bed_table1. Use it only if you know what you are doing. Typically useful for multiple
correlations testing so that the NULL distribution can be calculated only once}

\item{resampled_bed_regions2_list}{supply your own resampled bed_table2. Use it only if you know what you are doing. Typically useful for multiple
correlations testing so that the NULL distribution can be calculated only once}

\item{nparallel}{how many parallel processes to use when running the resampling}

\item{verbose}{set to FALSE to suppress the info messages. Warning and error messages will still be shown}
}
\value{
object with correlation statistics and additional data
}
\description{
Given two sets of regions, calculate the overlap between the regions in the first
set and the regions in the second set (see the function intersectBed).
If nsamples is greater than zero, then the regions are resampled
randomly so that a NULL distribution of overlaps can be obtained. Notice that
the the bigger nsamples is, the more accurate the p-value for rejecting the NULL
hypothesis will be. A returned p-value of 0 just means that the p-value reached
the limit of decimal numbers that can be obtained and should be interpreted as
"< 1/nsamples".
}
