% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{getIRD}
\alias{getIRD}
\title{Compute inter-region distance of a set of bed regions}
\usage{
getIRD(bed_table)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end}
}
\value{
left, right and average inter-region distance
}
\description{
This function returns both the left and right IRD of each given region
as well as the average IRD. Regions must be non-overlapping
}
