% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersectBed}
\alias{intersectBed}
\title{Intersect two sets of bed regions}
\usage{
intersectBed(bed_table1, bed_table2)
}
\arguments{
\item{bed_table1}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique.}

\item{bed_table2}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique.}
}
\value{
object with details intersection statistics
}
\description{
Given two tables of bed regions, find the regions from the first table that
overlap the regions in the second table, and conversely the regions in the second
table that overlap the regions in the first table.
If the regions have classes, then find how many regions in the first table for each
class of regions are contained in each region or class of regions in the second table,
and find how many regions in the second table for each class of regions contain each
region or class of regions in the second table.
}
