% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersectPositionsAndBedRegions}
\alias{intersectPositionsAndBedRegions}
\title{Intersect positions with bed regions}
\usage{
intersectPositionsAndBedRegions(positions, bed_table, verbose = TRUE)
}
\arguments{
\item{positions}{data frame containing positions, with required columns chr, position, id and optionally class. Value in the id column must be unique}

\item{bed_table}{data frame containing bed regions, with required columns chr, start, end, id and optionally class. Value in the id column must be unique}

\item{verbose}{set to FALSE to suppress the info messages. Warning and error messages will still be shown}
}
\value{
object with details intersection statistics
}
\description{
Given a table of positions and a table of bed regions, find the positions that
are contained in the regions, and conversely the regions that overlap given positions.
If positions and/or bed regions have classes, then find how many positions for each
class of positions are contained in each region or class of regions, and find how many regions for
each class of regions contain each position or each class of positions.
This function allows for overlapping bed regions. In practice, if there are
overlapping bed regions, the regions in bed_table will be assigned to a minimum number of
sets such that each set contains non-overlapping regions. The function intersectPositionsAndBedRegions_nonOverlapping
will then be used on the non-overlapping sets separately, and the results merged.
}
