% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mergeAdjacentBedRegions}
\alias{mergeAdjacentBedRegions}
\title{Merge adjacent bed regions}
\usage{
mergeAdjacentBedRegions(bed_table, aggregateSignalMode = "sum")
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end, id, and optionally signal}

\item{aggregateSignalMode}{when merging bed regions that have a signal column,
options for aggregating signal are sum, mean, or weightedmean (weighted w.r.t size)}
}
\value{
updated bed_table with merged adjacent regions
}
\description{
Given a table with bed regions, merge regions that are adjacent and return
the updated table. The bed regions should not be overlapping. If there are
overlapping bed regions in the bed_table, consider running breakDownOverlappingBedRegions
to obtain non-overlapping segments and then run this function to merge the
adjacent segments. For example, given region id A, start 1, end 10, and region id B,
start 11, end 15, the function returns the merged region with id "A;B", start 1, end 15.
}
