% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plotBedSignalRegion}
\alias{plotBedSignalRegion}
\title{Plot signal of bed regions}
\usage{
plotBedSignalRegion(
  bed_table,
  bed_table2 = NULL,
  fileout = NULL,
  pchr,
  pstart,
  pend,
  signalColour = "black",
  signalColour2 = "grey",
  highlightRegions = NULL,
  highlightRegionsColours = NULL,
  highlightPositions = NULL,
  highlightPositionsColours = NULL,
  plotGenes = TRUE,
  proteinCodingOnly = TRUE,
  genomev = "hg19",
  main = "",
  lwd = 1.5,
  cexlabels = 1,
  ylabel = "signal",
  ylabel2 = "signal"
)
}
\arguments{
\item{bed_table}{data frame with required columns: chr, start, end, signal}

\item{bed_table2}{data frame with required columns: chr, start, end, signal}

\item{fileout}{name of the output file for the plot, use .pdf extension}

\item{pchr}{chromosome name of the region to plot}

\item{pstart}{left boundary position of the genomic region to plot}

\item{pend}{right boundary position of the genomic region to plot}

\item{signalColour}{colour of the signal line}

\item{signalColour2}{colour of the second signal line from bed_table2}

\item{highlightRegions}{highlightRegions is a list of bed tables (chr, start, end)}

\item{highlightRegionsColours}{highlightRegionsColours is a list of colours.
The names of the list need to match the names of the highlightRegions list}

\item{highlightPositions}{highlightPositions is a list of position tables (chr, position, optional text)}

\item{highlightPositionsColours}{highlightPositionsColours is a list of colours.
The names of the list need to match the names of the highlightPositions list}

\item{plotGenes}{if TRUE the coding genes in the regions will be plotted on top}

\item{proteinCodingOnly}{if TRUE then plot only protein coding genes}

\item{genomev}{hg19 or hg38}

\item{main}{title of the plot}

\item{lwd}{line width of the plot}

\item{cexlabels}{scaling parameter for the labels}

\item{ylabel}{ylabel for the bed_table signal}

\item{ylabel2}{ylabel for the bed_table2 signal}
}
\description{
Given a table of bed regions with the signal column, plot the signal values
across a given region. Signal for segments not included in the given bed regions
will be set to zero. A second bed table can be specified, and its signal will be
plotted along with the signal of the first bed table.
}
