% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sortChroms}
\alias{sortChroms}
\title{Sort chromosomes}
\usage{
sortChroms(chroms, decreasing = FALSE)
}
\arguments{
\item{chroms}{characters vector with chromosome names, chr prefix is optional, so, for example, "1" and "chr1" are both acceptable}

\item{decreasing}{if TRUE chromosomes are sorted in decreasing order from higher chromosome number}
}
\value{
ordered chromosome names vector
}
\description{
Order a list of chromosome names.
}
